requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
openpyxl==3.1.5
httpx==0.27.2
google-api-python-client==2.153.0
google-auth==2.36.0
python-multipart==0.0.12
aiofiles==24.1.0
apscheduler==3.10.4
python-dotenv==1.0.1
